pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "secret-scanner"
version = "0.1.0"
description = "A CLI tool to detect leaked secrets in Git repositories"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    "typer[all]>=0.12",
    "rich>=13",
    "pydantic>=2",
    "pyyaml>=6",
]

[project.scripts]
secret-scanner = "secret_scanner.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-cov>=5",
    "ruff>=0.4",
    "mypy>=1.10",
    "types-PyYAML>=6",
]

# Map the 'secret_scanner' package to the flat src/ directory.
# secret_scanner       → src/
# secret_scanner.engines → src/engines/
# secret_scanner.output  → src/output/
[tool.setuptools]
packages = ["secret_scanner", "secret_scanner.engines", "secret_scanner.output"]

[tool.setuptools.package-dir]
secret_scanner = "src"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "C4", "PTH"]
ignore = ["B008"]  # typer uses function calls in defaults

[tool.ruff.lint.isort]
known-first-party = ["secret_scanner"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
exclude = ["tests/"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--tb=short -q"

[tool.coverage.run]
source = ["secret_scanner"]
branch = true
omit = ["*/tests/*"]

[tool.coverage.report]
show_missing = true
fail_under = 80
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
